Trace a function and its children
The tracing starts if depth is zero and the current function is the root
function the user wants to trace.

When depth is greater than zero, it means the current function is a children of
a root graph function, and we should continue tracing.

When depth reaches zero again, the root function has exited and the function is
not traced.

The root function is hardcoded and only counting is done for now. Using the
vfs_fstat as a test, the module produces this output:

[958970.986965] register=ffffffff81143640 unregister=ffffffff811434e0 root_func=ffffffff812130c0
[958971.005286] register fgraph hooks ret=0
[958971.005288] lttng-fgraph loaded
[958989.548535] unregister fgraph hooks
[958989.548539] lttng-fgraph removed
[958989.548540] entries=1206 returns=1206

The number of entries is equal to the number of returns and no crash occured yet.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>