Add tracepoint with recursion protection
The tracepoint may call other kernel functions that may be traced. Protect from
infinite recursion using the mechanism in ftrace. It uses the same
current->trace_recursion field. It *should* be safe, because the function graph
hook is called after the main ftrace hook, but more investigation is required.

Notice that without recursion protection to both entry and return tracepoints,
it crashes violently. panic=0 in the cmdline is ignored and the machine reboots
immediately. I tried kdump and it does not produce a dump. I cannot use the
tracer itself, so I found the solution using trial and error.

Of course, the code is ugly, but it's only for a proof of concept.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>